refactor(sanity): move number input into directory
This paves the way for adding a new number input implementation based
on Portable Text Editor.
