+++ packages/sanity/src/core/form/inputs/index.ts
@@ -12,7 +12,7 @@ export * from './BooleanInput'
 export * from './DateInputs'
 export * from './EmailInput'
 export * from './files/types'
-export * from './NumberInput'
+export * from './NumberInput/NumberInput'
 export * from './ObjectInput'
 export * from './PortableText/PortableTextInput'
 export {PortableTextInput as BlockEditor} from './PortableText/PortableTextInput'
