+++ packages/sanity/src/core/form/studio/inputResolver/defaultInputs.ts
@@ -5,7 +5,7 @@ import {UniversalArrayInput} from '../../inputs/arrays/UniversalArrayInput'
 import {BooleanInput} from '../../inputs/BooleanInput'
 import {DateInput, DateTimeInput} from '../../inputs/DateInputs'
 import {EmailInput} from '../../inputs/EmailInput'
-import {NumberInput} from '../../inputs/NumberInput'
+import {NumberInput} from '../../inputs/NumberInput/NumberInput'
 import {ObjectInput} from '../../inputs/ObjectInput'
 import {SlugInput} from '../../inputs/Slug/SlugInput'
 import {StringInput} from '../../inputs/StringInput/StringInput'
